backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
openpyxl==3.1.2
mercadopago==2.2.2
python-dotenv==1.0.1
google-generativeai==0.7.2
